Fix localhost with port number response matching
The Match() attempts to find a match between the requestURL and all
registered responses on the server.  Prior to the match it prepares the
requestURL by removing the current origin if the requestURL is the same,
therefore providing the flexibility to match pathnames only.  This
normal process breaks if the requestURL has a specified port number in
the origin.  This fix makes sure port numbers are considered in the
default location and the full origin (all the way to the first pathname
forward slash of the URL) is replaced.

This commit includes a test case as well to demonstrate this resolution.

Resolves: #161