Fix respondWith registration of origins with port numbers
The respondWith() function relies on the 'path-to-regexp' module to
create a regular expression for a given url string.  When providing an
URL origin with a port number, 'path-to-regexp' interprets a colon ':'
to be a delimiter for a named group in the regular expression.  This
generates a regular expression that accepts anything prior to a forward
slash '/' which ends the grouping.  This means the same origins with
different port numbers are treated as if they are the same path.  This
commit resolves this issue by escaping the colon if a port number is
detected in the origin and then passing it on to the 'path-to-regexp'
module.

This commit fixes only strings that are passed as the URL parameter
which are then converted to Regexp in the function.  If there are other
colon's in the URL, they will not be escaped.  This should not happen as
symbols in a URL's pathname's should be urlencoded (%+decimal code).

This commit also includes a test case that demonstrates the confliction
of URL's with different port numbers but the same path values.

Resolves: #162