codejedi365-fix-readystatechange-flow > resolved
\### Rationale
The code was using fall through logic which caused the extra
readystatechange event to be thrown after the loadend event.  The logic
was there to delinate from a finished XHR request and throw the loadend
but not to stop processing. Note that a readystatechange did have to be
added prior to the loadend otherwise some requests would not resolve
and it was noticeable through other unit-tests. This commit has all
passing tests to include the one designed to replicate the issue of

Resolves #165