fix respond() & readystatechange process flow (#184)
* codejedi365-fix-readystatechange-flow > add test

### Rationale
Added issue replication test for issue #165 where readystatechange
event is occuring after loadend event occurs.

* codejedi365-fix-readystatechange-flow > resolved

### Rationale
The code was using fall through logic which caused the extra
readystatechange event to be thrown after the loadend event.  The logic
was there to delinate from a finished XHR request and throw the loadend
but not to stop processing. Note that a readystatechange did have to be
added prior to the loadend otherwise some requests would not resolve
and it was noticeable through other unit-tests. This commit has all
passing tests to include the one designed to replicate the issue of

Resolves #165

* codejedi365-fix-readystatechange-flow > improve codestyle

### Rationale
Ensure logic is clear by avoiding empty return statements and fall-through logic